+++ blockchain-core/src/risk/pulseguard/ingest/finalized_tail.rs
@@ -8,9 +8,9 @@ pub async fn stream_finalized(tx: Sender<RiskEvent>, blocks: Vec<Block>) {
     for b in blocks {
         for t in &b.transactions {
             let (to_addr, amount_u128) = match t {
-                Transaction::Transfer(tr) => (tr.to.clone(), tr.amount as u128),
-                Transaction::Call(c) => (c.to.clone(), c.value as u128),
-                Transaction::Stake(s) => (s.validator.clone(), s.amount as u128),
+                Transaction::Transfer(tr) => (tr.to.clone(), tr.amount),
+                Transaction::Call(c) => (c.to.clone(), c.value),
+                Transaction::Stake(s) => (s.validator.clone(), s.amount),
                 Transaction::Deploy(_) => ("".to_string(), 0u128),
                 Transaction::AIRequest(_) => ("".to_string(), 0u128),
             };
